refactor: move get functionality from iroh/main into  iroh-bytes/get/io
